requirements.txt
click==8.0.0
dnspython==2.1.0
Flask==2.0.0
Flask-JWT-Extended==4.2.1
gunicorn==20.1.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.17.1
Werkzeug==2.0.0
